hlsdemux: update playlist immediately when resuming after pause
When going from PLAYING to PAUSED and then back to PLAYING we don't
want to schedule an update of the playlist using the usual schedule
function as this would just add fragment duration to the absolute
time we stored in the past which can lead to random behavior. Instead
we update the playlist immediately when resuming and then schedule
updates normally after that.