Fix bug with attrib/value with no ending ',' and DOS linefeeds.
The original function handled attribute parsing poorly. It wasn't
clear if it handled attrib/value lacking a trailing ','.
Likewise, it did not handle DOS /r/n characters tacked on to the
end of the value of a terminating attribute. The new function
strips whitespace from the end of the value.