h264parse: swap PAR priority
Before it was:
1 - container
2 - codec.
Now it's
1 - codec
2 - container

The reason to change is that Ligada Test
org.hbbtv_00002320 fails, having correct
PAR in the codec and incorrect one in the
container. In fact such priority order
should be specified in HTML5 standard.
For now all the PAR related tests are passing,
that's enough for us.