amc: fix occasional deadlock on seeking
It may happen that code will loop in
gst_amc_video_dec_handle_frame, if
android don't have free input buffers,
because they all are in use.
For this case on FLUSH_START we must
avoid the deadlock, and tell to
gst_amc_video_dec_handle_frame that
we are flushing.
Base class is resetting decoder on
FLUSH_STOP.