openslessink: tell basesink to sync with ts
BaseAudioSink sets BaseSink not to wait for
next buffer's timestamp, and just believes
that audio stream is a solid stream of raw
audio samples.
However we have found a file that has jumps
in audio buffers timestamps (probably the
gap should be treated as silence), so
if we don't wait for that timestamps,
audio pipeline consumes data faster then
video one, and it makes multiqueue
stuck at some point.