Fix adaptive mode
Adaptive playback means that the decoder can change the frame size on the fly,
so that no destruction is necessary to reconfigure.

Now if adaptive is enabled, we restart the amc codec only if format changed
more than the frame size.

The maximum frame size was being set with the current frame size. It needs to
be the maximum frame size to be decoded, but in dash we cannot know beforehand.
Now it is left unset, and verified to work with hikeen board. Documentation is
not clear if this is allowed, although leads to think so by tagging this as
optional and a hint for the codec to optimize resource allocation. In case
we find issues with other codec implementations, we'd need to set it from the
maximum size supported by the codec, at least when the maximum size cannot
be known.

This fixed the black gaps when the stream format changed in adaptive mode. It
still can happen if adaptive is not supported (unlikely in our targets), or when
the encoding changes. To fix those cases we cannot invalidate pushed buffers as
we are doing now. But that would be a more complex change that does not need to
be addressed urgently because this one solved the issue reported.