amc: class_init for dynamic subtypes of AmcVideoDec
Each subtype of AmcVideoDec represents an android codec and MIME
type combination; which represents an entry in media_codecs.xml

Each of these subtypes can have a different set of properties
which represent one of the codec features in media_codecs.xml.
These features are now set in the  class_init function for these
dynamic subtypes. The class_init function for these is called
gst_amc_video_dec_dynamic_class_init.

The creation of the description string which can be used to filter
the facotry lists by features was moved to base_init function.