amc: record if feature is required / not optional
If a feature is required it means that it is not possible to use
the factory without the feature enabled. Marking this in the
factory description allows it to be filtered out if the feature
is not required / supported. This filtering is done in
gstelementfactory.c.

Also removed features which are not currently being used, these
can be added in the future when implemented.